available_ppc {bayesplot}	R Documentation

Get or view the names of available plotting functions

Description

Get or view the names of available plotting functions

Usage

available_ppc(pattern = NULL, fixed = FALSE, invert = FALSE)

available_mcmc(pattern = NULL, fixed = FALSE, invert = FALSE)

Arguments

pattern, fixed, invert
Passed to base::grep().

Value

A possibly empty character vector of function names with several additional attributes (for use by a custom print method). If pattern is missing then the returned object contains the names of all available plotting functions in the MCMC or PPC module, depending on which function is called. If pattern is specified then a subset of function names is returned.

Examples

available_mcmc()
available_mcmc("nuts")
available_mcmc("rhat|neff")
available_ppc("grouped")
available_ppc("grouped", invert = TRUE)

[Package bayesplot version 1.7.2 Index]