Get or view the names of available plotting functions

Get or view the names of available plotting functions

available_ppc(pattern, fixed = FALSE, invert = FALSE)

available_mcmc(pattern, fixed = FALSE, invert = FALSE)

Arguments

pattern, fixed, invert
Passed to grep.

Value

A possibly empty character vector of function names with several additional attributes (for use by a custom print method). If pattern is missing then the returned object contains the names of all available plotting functions in the MCMC or PPC module, depending on which function is called. If pattern is specified then a subset of function names is returned.

Examples

available_mcmc()

#> bayesplot MCMC module:
#>   mcmc_acf
#>   mcmc_acf_bar
#>   mcmc_areas
#>   mcmc_areas_data
#>   mcmc_areas_ridges
#>   mcmc_areas_ridges_data
#>   mcmc_combo
#>   mcmc_dens
#>   mcmc_dens_chains
#>   mcmc_dens_chains_data
#>   mcmc_dens_overlay
#>   mcmc_hex
#>   mcmc_hist
#>   mcmc_hist_by_chain
#>   mcmc_intervals
#>   mcmc_intervals_data
#>   mcmc_neff
#>   mcmc_neff_data
#>   mcmc_neff_hist
#>   mcmc_nuts_acceptance
#>   mcmc_nuts_divergence
#>   mcmc_nuts_energy
#>   mcmc_nuts_stepsize
#>   mcmc_nuts_treedepth
#>   mcmc_pairs
#>   mcmc_parcoord
#>   mcmc_parcoord_data
#>   mcmc_recover_hist
#>   mcmc_recover_intervals
#>   mcmc_recover_scatter
#>   mcmc_rhat
#>   mcmc_rhat_data
#>   mcmc_rhat_hist
#>   mcmc_scatter
#>   mcmc_trace
#>   mcmc_trace_highlight
#>   mcmc_violin

available_mcmc("nuts")

#> bayesplot MCMC module:
#> (matching pattern 'nuts') 
#>   mcmc_nuts_acceptance
#>   mcmc_nuts_divergence
#>   mcmc_nuts_energy
#>   mcmc_nuts_stepsize
#>   mcmc_nuts_treedepth

available_mcmc("rhat|neff")

#> bayesplot MCMC module:
#> (matching pattern 'rhat|neff') 
#>   mcmc_neff
#>   mcmc_neff_data
#>   mcmc_neff_hist
#>   mcmc_rhat
#>   mcmc_rhat_data
#>   mcmc_rhat_hist

available_ppc("grouped")

#> bayesplot PPC module:
#> (matching pattern 'grouped') 
#>   ppc_bars_grouped
#>   ppc_error_hist_grouped
#>   ppc_freqpoly_grouped
#>   ppc_intervals_grouped
#>   ppc_ribbon_grouped
#>   ppc_scatter_avg_grouped
#>   ppc_stat_freqpoly_grouped
#>   ppc_stat_grouped
#>   ppc_violin_grouped

available_ppc("grouped", invert = TRUE)

#> bayesplot PPC module:
#> (excluding pattern 'grouped') 
#>   ppc_bars
#>   ppc_bars_yrep_data
#>   ppc_boxplot
#>   ppc_data
#>   ppc_dens
#>   ppc_dens_overlay
#>   ppc_ecdf_overlay
#>   ppc_error_binned
#>   ppc_error_hist
#>   ppc_error_scatter
#>   ppc_error_scatter_avg
#>   ppc_error_scatter_avg_vs_x
#>   ppc_freqpoly
#>   ppc_group_data
#>   ppc_hist
#>   ppc_intervals
#>   ppc_intervals_data
#>   ppc_loo_intervals
#>   ppc_loo_pit
#>   ppc_loo_pit_overlay
#>   ppc_loo_pit_qq
#>   ppc_loo_ribbon
#>   ppc_ribbon
#>   ppc_ribbon_data
#>   ppc_rootogram
#>   ppc_scatter
#>   ppc_scatter_avg
#>   ppc_stat
#>   ppc_stat_2d